% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{ucbthesis}
\alias{ucbthesis}
\alias{ucbthesis-package}
\title{UC Berkeley Graduate Division Thesis Templates}
\description{
This package contains latex, knitr and R Markdown templates that
adhere to the UC Berkeley Graduate Division's thesis guidelines. The templates
are located in the \code{inst/} directory.
}

